update imports for basic discord bots
