+++ bots/discord_price_bot.py
@@ -12,7 +12,7 @@ if __name__ == "__main__":
         level=logging.INFO,
         handlers=[
             logging.handlers.RotatingFileHandler(
-                "discord_price_bot.log", maxBytes=util.ONE_HUNDRED_MEGABYTES, backupCount=1
+                "discord_price_bot.log", maxBytes=ONE_HUNDRED_MEGABYTES, backupCount=1
             ),
             logging.StreamHandler(),
         ],
