+++ bots/telegram_bot.py
@@ -91,7 +91,7 @@ if __name__ == "__main__":
         level=logging.INFO,
         handlers=[
             logging.handlers.RotatingFileHandler(
-                "telegram_bot.log", maxBytes=util.ONE_HUNDRED_MEGABYTES, backupCount=1
+                "telegram_bot.log", maxBytes=ONE_HUNDRED_MEGABYTES, backupCount=1
             ),
             logging.StreamHandler(),
         ],
