+++ bots/util.py
@@ -1,162 +1,11 @@
-from abc import abstractmethod
-import asyncio.exceptions
-from datetime import datetime, timedelta
-import discord
-from discord.ext import tasks, commands
-from enum import Enum
 import re
 import logging
-from opensea import OpenseaAPI
-import os
-import subprocess
 import sys
 import threading
 import time
-import websockets
 import re
 
-from constants.addresses import *
-from data_access.graphs import (
-    BeanSqlClient,
-    BeanstalkSqlClient,
-    BasinSqlClient,
-    SnapshotClient,
-    DAO_SNAPSHOT_NAME,
-)
 from data_access.eth_chain import *
-from data_access.util import *
-from data_access.etherscan import get_gas_base_fee
-from data_access.coin_gecko import get_token_price
-from data_access.eth_usd_oracle import *
-from tools.util import get_txn_receipt_or_wait
-
-# Strongly encourage Python 3.8+.
-# If not 3.8+ uncaught exceptions on threads will not be logged.
-MIN_PYTHON = (3, 8)
-if sys.version_info < MIN_PYTHON:
-    logging.critical(
-        "Python %s.%s or later is required for proper exception logging.\n" % MIN_PYTHON
-    )
-LOGGING_FORMAT_STR_SUFFIX = "%(levelname)s : %(asctime)s : %(message)s"
-LOGGING_FORMATTER = logging.Formatter(LOGGING_FORMAT_STR_SUFFIX)
-
-TIMESTAMP_KEY = "timestamp"
-# Discord server guild ID.
-# BEANSTALK_GUILD_ID = 880413392916054098
-ID_KEY = "id"
-# The duration of a season. Assumes that seasons align with Unix epoch.
-SEASON_DURATION = 3600  # seconds
-# How long to wait between discord preview bot updates.
-PREVIEW_CHECK_PERIOD = 4  # seconds
-# For all check periods there is a built in assumption that we will update at least once per
-# Ethereum block (~13.5 seconds).
-APPROX_BLOCK_TIME = 12  # seconds
-# How long to wait between peg checks.
-PEG_CHECK_PERIOD = APPROX_BLOCK_TIME  # seconds
-# How long to wait between checks for a sunrise when we expect a new season to begin.
-SUNRISE_CHECK_PERIOD = APPROX_BLOCK_TIME  # seconds
-# Rate at which to check chain for new Uniswap V2 pool interactions.
-POOL_CHECK_RATE = APPROX_BLOCK_TIME  # seconds
-# Rate at which to check for events on the Beanstalk contract.
-BEANSTALK_CHECK_RATE = APPROX_BLOCK_TIME  # seconds
-# How long to wait between checks for fert purchases.
-BARN_RAISE_CHECK_RATE = APPROX_BLOCK_TIME  # seconds
-# Bytes in 100 megabytes.
-ONE_HUNDRED_MEGABYTES = 100 * 1000000
-# Initial time to wait before reseting dead monitor.
-RESET_MONITOR_DELAY_INIT = 15  # seconds
-# Timestamp for deployment of Basin.
-BASIN_DEPLOY_EPOCH = 1692814103
-
-DISCORD_NICKNAME_LIMIT = 32
-
-class DiscordSidebarClient(discord.ext.commands.Bot):
-    def __init__(self, monitor, prod=False):
-        super().__init__(command_prefix=commands.when_mentioned_or("!"))
-
-        self.nickname = ""
-        self.last_nickname = ""
-        self.status_text = ""
-
-        # Try to avoid hitting Discord API rate limit when all bots starting together.
-        time.sleep(1.1)
-
-        # subclass of util.Monitor
-        self.monitor = monitor(self.set_nickname, self.set_status)
-        self.monitor.start()
-
-        # Ignore exceptions of this type and retry. Note that no logs will be generated.
-        # Ignore base class, because we always want to reconnect.
-        # https://discordpy.readthedocs.io/en/latest/api.html#discord.ClientUser.edit
-        # https://discordpy.readthedocs.io/en/latest/api.html#exceptions
-        self._update_naming.add_exception_type(discord.DiscordException)
-
-        # Start the price display task in the background.
-        self._update_naming.start()
-
-    def stop(self):
-        self.monitor.stop()
-
-    def set_nickname(self, text):
-        """Set bot server nickname."""
-        self.nickname = text
-
-    def set_status(self, text):
-        """Set bot custom status text."""
-        self.status_text = text
-
-    async def on_ready(self):
-        # Log the commit of this run.
-        logging.info(
-            "Git commit is "
-            + subprocess.check_output(
-                ["git", "rev-parse", "--short", "HEAD"],
-                cwd=os.path.dirname(os.path.realpath(__file__)),
-            )
-            .decode("ascii")
-            .strip()
-        )
-
-        self.user_id = self.user.id
-        # self.beanstalk_guild = self.get_guild(BEANSTALK_GUILD_ID)
-        # Guild IDs for all servers this bot is in.
-        self.current_guilds = []
-        for guild in self.guilds:
-            self.current_guilds.append(guild)
-            logging.info(f"Guild found: {guild.id}")
-
-    @tasks.loop(seconds=1, reconnect=True)
-    async def _update_naming(self):
-        if self.nickname:
-            await update_discord_bot_name(self.nickname, self)
-        self.nickname = ""
-        if self.status_text:
-            await self.change_presence(
-                activity=discord.Activity(type=discord.ActivityType.watching, name=self.status_text)
-            )
-            logging.info(f"Bot status changed to {self.status_text}")
-            self.status_text = ""
-
-    @_update_naming.before_loop
-    async def before__update_naming_loop(self):
-        """Wait until the bot logs in."""
-        await self.wait_until_ready()
-
-
-async def update_discord_bot_name(name, bot):
-    emoji_accent = holiday_emoji()
-    next_name = emoji_accent + name + emoji_accent
-    # Discord character limit
-    if len(next_name) > DISCORD_NICKNAME_LIMIT:
-        pruned_name = next_name[: DISCORD_NICKNAME_LIMIT - 3] + "..."
-        logging.info(f"Pruning nickname from {next_name} to {pruned_name}")
-        next_name = pruned_name
-    # Note(funderberker): Is this rate limited?s
-    for guild in bot.current_guilds:
-        logging.info(f"Attempting to set nickname in guild with id {guild.id}")
-        await guild.me.edit(nick=next_name)
-        logging.info(f"Bot nickname changed to {next_name} in guild with id {guild.id}")
-    return next_name
 
 class MsgHandler(logging.Handler):
     """A handler class which sends a message on a text channel."""
@@ -179,10 +28,6 @@ class MsgHandler(logging.Handler):
             self.message_function(msg)
         except Exception:
             self.handleError(record)
-
-    def __repr__(self):
-        level = getLevelName(self.level)
-        return "<%s %s (%s)>" % (self.__class__.__name__, self.baseFilename, level)
 
 
 def event_in_logs(name, event_logs):
@@ -403,19 +248,3 @@ def log_exceptions(exc_type, exc_value, exc_traceback):
 
 def configure_main_thread_exception_logging():
     sys.excepthook = log_exceptions
-
-
-if __name__ == "__main__":
-    """Quick test and demonstrate functionality."""
-    logging.basicConfig(level=logging.INFO)
-
-    sunrise_monitor = SeasonsMonitor(print)
-    sunrise_monitor.start()
-
-    try:
-        while True:
-            time.sleep(1)
-    except KeyboardInterrupt:
-        pass
-
-    sunrise_monitor.stop()
