+++ constants/addresses.py
@@ -1,3 +1,4 @@
+# EVM addresses
 NULL_ADDR = "0x0000000000000000000000000000000000000000"
 WRAPPED_ETH = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2"
 BEAN_ADDR = "0xBEA0000029AD1c77D3d5D23Ba2D8893dB9d1Efab"
