+++ constants/channels.py
@@ -1,0 +1,22 @@
+# Channel ids for posting messages to discord/telegram channels etc
+
+BS_DISCORD_CHANNEL_ID_PEG_CROSSES = 911338190198169710
+BS_DISCORD_CHANNEL_ID_SEASONS = 911338078080221215
+BS_DISCORD_CHANNEL_ID_POOL = 915372733758603284
+BS_DISCORD_CHANNEL_ID_BEANSTALK = 918240659914227713
+BS_DISCORD_CHANNEL_ID_MARKET = 940729085095723069
+BS_DISCORD_CHANNEL_ID_REPORT = 943711736391933972
+BS_DISCORD_CHANNEL_ID_BARN_RAISE = 969594841455558717
+BS_DISCORD_CHANNEL_ID_TEST_BOT = 908035718859874374
+
+BS_TELEGRAM_FWD_CHAT_ID_TEST = "-1001655547288"
+BS_TELEGRAM_FWD_CHAT_ID_PRODUCTION = "@beanstalkUSD"
+
+BS_DISCORD_CHANNEL_ID_ANNOUNCEMENTS = 880500642546851850
+BS_DISCORD_CHANNEL_ID_WEEKLY_UPDATES = 1025448845594861718
+
+DEX_DISCORD_CHANNEL_ID_TEST = 1117767077420339220
+DEX_DISCORD_CHANNEL_ID_DAILY = 1117768293625888868
+DEX_DISCORD_CHANNEL_ID_BEAN_ETH = 1117768417861193819
+DEX_DISCORD_CHANNEL_ID_BEAN_WSTETH = 1262891110439587890
+DEX_DISCORD_CHANNEL_ID_WELLS_OTHER = 1257633554427543633
