+++ monitors/curve.py
@@ -1,3 +1,13 @@
+from data_access.coin_gecko import get_token_price
+
+from bots.util import *
+from monitors.monitor import Monitor
+from data_access.eth_chain import *
+from data_access.graphs import *
+from data_access.util import *
+from constants.addresses import *
+from constants.config import *
+
 class CurvePoolMonitor(Monitor):
     """Monitor a Curve pool for events."""
