+++ monitors/preview/basin.py
@@ -1,3 +1,11 @@
+from bots.util import *
+from monitors.preview.preview import PreviewMonitor
+from data_access.eth_chain import *
+from data_access.graphs import *
+from data_access.util import *
+from constants.addresses import *
+from constants.config import *
+
 class BasinStatusPreviewMonitor(PreviewMonitor):
     """Monitor data that offers view into current Basin token status via discord nickname/status.
