+++ monitors/preview/eth.py
@@ -1,3 +1,14 @@
+from data_access.etherscan import get_gas_base_fee
+from data_access.eth_usd_oracle import get_twa_eth_price
+
+from bots.util import *
+from monitors.preview.preview import PreviewMonitor
+from data_access.eth_chain import *
+from data_access.graphs import *
+from data_access.util import *
+from constants.addresses import *
+from constants.config import *
+
 class EthPreviewMonitor(PreviewMonitor):
     """Monitor data that offers a view into Eth mainnet."""
