+++ monitors/seasons.py
@@ -1,3 +1,11 @@
+from monitors.monitor import Monitor
+from data_access.eth_chain import *
+from data_access.graphs import *
+from data_access.util import *
+from data_access.eth_usd_oracle import *
+from constants.addresses import *
+from constants.config import *
+
 class SeasonsMonitor(Monitor):
     def __init__(
         self, message_function, short_msgs=False, channel_to_wallets=None, prod=False, dry_run=False
